requirements.txt
opencv-python
Pillow
scikit-learn
numpy
matplotlib
